pyproject.toml
[project]
name = "ticket-price-predictor"
version = "0.1.0"
description = "ML system for predicting ticket price changes at the seat-zone level"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "playwright>=1.40.0",
    "playwright-stealth>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyarrow>=15.0.0",
    "selectolax>=0.3.0",
    "scikit-learn>=1.4.0",
    "lightgbm>=4.3.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "joblib>=1.3.0",
    "ytmusicapi>=1.5.0",
    "optuna>=4.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "python-dotenv>=1.0.0",
]
ml = [
    "optuna>=3.5.0",
    "shap>=0.44.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]
neural = [
    "pytorch-tabular>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ticket_price_predictor"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["ticket_price_predictor"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
disable_error_code = ["import-untyped"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["pyarrow", "pyarrow.*"]
ignore_missing_imports = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = ["optuna", "optuna.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["ytmusicapi", "ytmusicapi.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["pytorch_tabular", "pytorch_tabular.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --cov=src/ticket_price_predictor --cov-report=term-missing"

[tool.coverage.run]
source = ["src/ticket_price_predictor"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]
